v6.2.2

Conflicts:
	core/src/.gitignore
	core/src/.htaccess
	core/src/conf/bootstrap_repositories.php
	core/src/data/.htaccess
	core/src/plugins/access.fs/class.fsAccessDriver.php
	core/src/plugins/access.fs/fsActions.xml
	core/src/plugins/access.ftp/manifest.xml
	core/src/plugins/gui.ajax/res/themes/vision/css/allz.css

+++ core/src/plugins/uploader.jumploader/class.JumploaderProcessor.php
@@ -35,6 +35,7 @@ class JumploaderProcessor extends AJXP_Plugin
      */
     private static $skipDecoding = false;
     private static $remote = false;
+    private static $wrapperIsRemote = false;
     private static $partitions = array();
 
     public function preProcess($action, &$httpVars, &$fileVars)
@@ -47,16 +48,18 @@ class JumploaderProcessor extends AJXP_Plugin
             self::$remote = true;
         }
 
-        if ($repository->detectStreamWrapper(false)) {
-            $plugin = AJXP_PluginsService::findPlugin("access", $repository->getAccessType());
-            $streamData = $plugin->detectStreamWrapper(true);
-            if ($streamData["protocol"] == "ajxp.ftp" || $streamData["protocol"]=="ajxp.remotefs") {
+        if ($repository->detectStreamWrapper(true)) {
+            $wrapperName = AJXP_MetaStreamWrapper::actualRepositoryWrapperClass($repository->getId());
+            if ($wrapperName == "ajxp.ftp" || $wrapperName == "ajxp.remotefs") {
                 $this->logDebug("Skip decoding");
                 self::$skipDecoding = true;
             }
+            $this->logDebug("Stream is",$wrapperName);
+            self::$wrapperIsRemote = call_user_func(array($wrapperName, "isRemote"));
         }
         $this->logDebug("Jumploader HttpVars", $httpVars);
         $this->logDebug("Jumploader FileVars", $fileVars);
+
 
         $httpVars["dir"] = base64_decode(str_replace(" ","+",$httpVars["dir"]));
         $index = $httpVars["partitionIndex"];
@@ -64,12 +67,12 @@ class JumploaderProcessor extends AJXP_Plugin
 
         /* if fileId is not set, request for cross-session resume (only if the protocol is not ftp)*/
         if (!isSet($httpVars["fileId"])) {
-            AJXP_LOGGER::debug("Cross-Session Resume request");
+            $this->logDebug("Trying Cross-Session Resume request");
 
-            $plugin = AJXP_PluginsService::findPlugin("access", $repository->getAccessType());
-            $streamData = $plugin->detectStreamWrapper(true);
             $dir = AJXP_Utils::decodeSecureMagic($httpVars["dir"]);
-            $destStreamURL = $streamData["protocol"]."://".$repository->getId().$dir;
+            $repository->detectStreamWrapper(true);
+            $context = new UserSelection($repository);
+            $destStreamURL = $context->currentBaseUrl().$dir;
             $fileHash = md5($httpVars["fileName"]);
 
             if (!self::$remote) {
@@ -85,15 +88,17 @@ class JumploaderProcessor extends AJXP_Plugin
                             $resumeFileId = $explodedSubPathName[1];
                             $resumeIndexes[] = $explodedSubPathName[2];
 
-                            AJXP_LOGGER :: debug("Current Index: " . $explodedSubPathName[2]);
+                            $this->logDebug("Current Index: " . $explodedSubPathName[2]);
                         }
                     }
                     $it->next();
                 }
 
                 /* no valid temp file found. return. */
-                if (empty ($resumeIndexes))
+                if (empty ($resumeIndexes)){
+                    $this->logDebug("No Cross-Session Resume request");
                     return;
+                }
 
                 AJXP_LOGGER :: debug("ResumeFileID: " . $resumeFileId);
                 AJXP_LOGGER :: debug("Max Resume Index: " . max($resumeIndexes));
@@ -101,7 +106,7 @@ class JumploaderProcessor extends AJXP_Plugin
                 AJXP_LOGGER :: debug("Next Resume Index: " . $nextResumeIndex);
 
                 if (isSet($resumeFileId)) {
-                    AJXP_LOGGER::debug("ResumeFileId is set. Returning values: fileId: " . $resumeFileId . ", partitionIndex: " . $nextResumeIndex);
+                    $this->logDebug("ResumeFileId is set. Returning values: fileId: " . $resumeFileId . ", partitionIndex: " . $nextResumeIndex);
                     $httpVars["resumeFileId"] = $resumeFileId;
                     $httpVars["resumePartitionIndex"] = $nextResumeIndex;
                 }
@@ -111,17 +116,33 @@ class JumploaderProcessor extends AJXP_Plugin
 
         /* if the file has to be partitioned */
         if (isSet($httpVars["partitionCount"]) && intval($httpVars["partitionCount"]) > 1) {
-            AJXP_LOGGER::debug("Partitioned upload");
+            $this->logDebug("Partitioned upload");
             $fileId = $httpVars["fileId"];
             $fileHash = md5($realName);
 
             /* In order to enable cross-session resume, temp files must not depend on session.
              * Now named after and md5() of the original file name.
              */
-            AJXP_LOGGER::debug("Filename: " . $realName . ", File hash: " . $fileHash);
+            $this->logDebug("Filename: " . $realName . ", File hash: " . $fileHash);
             $fileVars["userfile_0"]["name"] = "$fileHash.$fileId.$index";
             $httpVars["lastPartition"] = false;
+        }else{
+            /*
+             * If we wan to upload a folderUpload to folderServer
+             * Temporarily,put all files in this folder to folderServer.
+             * But a same file name may be existed in folderServer,
+             * this can cause error of uploading.
+             *
+             * We rename this file by his relativePath. At the postProcess session, we will use this name
+             * to copy to right location
+             *
+             */
+            $file_tmp_md5 = md5($httpVars["relativePath"]);
+            $fileVars["userfile_0"]["name"] = $file_tmp_md5;
         }
+
+
+
         /* if we received the last partition */
         if (intval($index) == intval($httpVars["partitionCount"])-1) {
             $httpVars["lastPartition"] = true;
@@ -152,9 +173,9 @@ class JumploaderProcessor extends AJXP_Plugin
                 /* we get the stream url (where all the partitions have been uploaded so far) */
                 $repository = ConfService::getRepository();
                 $dir = AJXP_Utils::decodeSecureMagic($httpVars["dir"]);
-                $plugin = AJXP_PluginsService::findPlugin("access", $repository->getAccessType());
-                $streamData = $plugin->detectStreamWrapper(true);
-                $destStreamURL = $streamData["protocol"]."://".$repository->getId().$dir."/";
+                $repository->detectStreamWrapper(true);
+                $context = new UserSelection($repository);
+                $destStreamURL = $context->currentBaseUrl().$dir."/";
 
                 if ($httpVars["partitionCount"] > 1) {
                     /* we fetch the information that help us to construct the temp files name */
@@ -188,15 +209,16 @@ class JumploaderProcessor extends AJXP_Plugin
         }
 
         if ($httpVars["lastPartition"]) {
-            $plugin = AJXP_PluginsService::findPlugin("access", $repository->getAccessType());
-            $streamData = $plugin->detectStreamWrapper(true);
             $dir = AJXP_Utils::decodeSecureMagic($httpVars["dir"]);
-            $destStreamURL = $streamData["protocol"]."://".$repository->getId().$dir."/";
+            $repository->detectStreamWrapper(true);
+            $context = new UserSelection($repository);
+            $destStreamURL = $context->currentBaseUrl().$dir."/";
 
             /* we check if the current file has a relative path (aka we want to upload an entire directory) */
-            AJXP_LOGGER::debug("Now dispatching relativePath dest:", $httpVars["relativePath"]);
+            $this->logDebug("Now dispatching relativePath dest:", $httpVars["relativePath"]);
             $subs = explode("/", $httpVars["relativePath"]);
             $userfile_name = array_pop($subs);
+
             $folderForbidden = false;
             $all_in_place = true;
             $partitions_length = 0;
@@ -204,6 +226,15 @@ class JumploaderProcessor extends AJXP_Plugin
             $fileHash = md5($userfile_name);
             $partitionCount = $httpVars["partitionCount"];
             $fileLength = $_POST["fileLength"];
+
+            /*
+             *
+             * Now, we supposed that access driver has already saved uploaded file in to
+             * folderServer with file name is md5 relativePath value.
+             * We try to copy this file to right location in recovery his name.
+             *
+             */
+            $userfile_name = md5($httpVars["relativePath"]);
 
             if (self::$remote) {
                 $partitions = array();
@@ -297,13 +328,15 @@ class JumploaderProcessor extends AJXP_Plugin
                     $chmodValue = $repository->getOption("CHMOD_VALUE");
                     if (isSet($chmodValue) && $chmodValue != "") {
                         $dirMode = octdec(ltrim($chmodValue, "0"));
-                        if ($dirMode & 0400) $dirMode |= 0100; // User is allowed to read, allow to list the directory
+                        if ($dirMode & 0400) $dirMode |= 0100; // Owner is allowed to read, allow to list the directory
                         if ($dirMode & 0040) $dirMode |= 0010; // Group is allowed to read, allow to list the directory
                         if ($dirMode & 0004) $dirMode |= 0001; // Other are allowed to read, allow to list the directory
                     }
+                    $url = $destStreamURL.$curDir."/".$dirname;
                     $old = umask(0);
-                    mkdir($destStreamURL.$curDir."/".$dirname, $dirMode);
+                    mkdir($url, $dirMode);
                     umask($old);
+                    AJXP_Controller::applyHook("node.change", array(null, new AJXP_Node($url), false));
                     $curDir .= "/".$dirname;
                 }
             }
@@ -313,11 +346,23 @@ class JumploaderProcessor extends AJXP_Plugin
                 $this->logDebug("Should now rebuild file!", $httpVars);
                 // Now move the final file to the right folder
                 // Currently the file is at the base of the current
-                AJXP_LOGGER::debug("PartitionRealName", $destStreamURL.$httpVars["partitionRealName"]);
+                $this->logDebug("PartitionRealName", $destStreamURL.$httpVars["partitionRealName"]);
+
+                // Get file by name (md5 value)
+                $relPath_md5 = AJXP_Utils::decodeSecureMagic(md5($httpVars["relativePath"]));
+
+                // original file name
                 $relPath = AJXP_Utils::decodeSecureMagic($httpVars["relativePath"]);
+
                 $target = $destStreamURL;
                 $target .= (self::$remote)? basename($relPath) : $relPath;
-                $current = $destStreamURL.basename($relPath);
+
+                /*
+                *   $current is uploaded file with md5 value as his name
+                *   we copy to $relPath and delete md5 file
+                */
+
+                $current = $destStreamURL.basename($relPath_md5);
 
                 if ($httpVars["partitionCount"] > 1) {
                     if (self::$remote) {
@@ -329,10 +374,12 @@ class JumploaderProcessor extends AJXP_Plugin
                             $currentFile = $partitions[$i];
                             $currentFileName = $currentFile["tmp_name"];
                             $part = fopen($currentFileName, "r");
-                            while (!feof($part)) {
-                                $length += fwrite($newDest, fread($part, 4096));
+                            if(is_resource($part)){
+                                while (!feof($part)) {
+                                    $length += fwrite($newDest, fread($part, 4096));
+                                }
+                                fclose($part);
                             }
-                            fclose($part);
                             unlink($currentFileName);
                         }
                         $newFile["type"] = $partitions[0]["type"];
@@ -343,15 +390,18 @@ class JumploaderProcessor extends AJXP_Plugin
                         $newFile["destination"] = $partitions[0]["destination"];
                         $newPartitions[] = $newFile;
                     } else {
-                        $newDest = fopen($destStreamURL.$httpVars["partitionRealName"], "w");
+                        $current = $destStreamURL.$httpVars["partitionRealName"];
+                        $newDest = fopen($current, "w");
                         $fileHash = md5($httpVars["partitionRealName"]);
 
                         for ($i = 0; $i < $httpVars["partitionCount"] ; $i++) {
                             $part = fopen($destStreamURL."$fileHash.$fileId.$i", "r");
-                            while (!feof($part)) {
-                                fwrite($newDest, fread($part, 4096));
+                            if(is_resource($part)){
+                                while (!feof($part)) {
+                                    fwrite($newDest, fread($part, 4096));
+                                }
+                                fclose($part);
                             }
-                            fclose($part);
                             unlink($destStreamURL."$fileHash.$fileId.$i");
                         }
 
@@ -359,15 +409,18 @@ class JumploaderProcessor extends AJXP_Plugin
                     fclose($newDest);
                 }
 
-                if (!self::$remote) {
-                    $err = copy($current, $target);
+                if (!self::$remote && (!self::$wrapperIsRemote || $relPath != $httpVars["partitionRealName"])) {
+                    if($current != $target){
+                        $err = copy($current, $target);
+                    }
+                    else $err = true;
                 } else {
                     for ($i=0, $count=count($newPartitions); $i<$count; $i++) {
                         $driver->storeFileToCopy($newPartitions[$i]);
                     }
                 }
 
-                if ($err !== false) {
+                if ($current != $target && $err !== false) {
                     if(!self::$remote) unlink($current);
                     AJXP_Controller::applyHook("node.change", array(null, new AJXP_Node($target), false));
                 } else if ($current == $target) {
@@ -375,7 +428,7 @@ class JumploaderProcessor extends AJXP_Plugin
                 }
             } else {
                 // Remove the file, as it should not have been uploaded!
-                if(!self::$remote) unlink($current);
+                //if(!self::$remote) unlink($current);
             }
         }
     }
